Make tests not flaky anymore
The tests would randomly fail because they run in multiple threads but
in the same process and the cwd is thus shared among threads.
Depending on the order in which things happened this could lead to a cd
in the middle of a test, leading to some "not found" errors.

It's just best to avoid relying on the cwd in tests, instead we just
pass the path where things are supposed to happen as a function
argument.